% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dplus.R
\name{calc_dplus}
\alias{calc_dplus}
\title{Calculate Dplus for Allele Abundances}
\usage{
calc_dplus(genpop)
}
\arguments{
\item{genpop}{A genpop object (from the adegenet package) containing an additional row "pop" containing the KBA in which the individual occurs.}
}
\value{
A numeric vector containing the Dplus values for each KBA.
}
\description{
This function calculates the Dplus diversity metric using the
\code{taxondive()} function from the vegan package.
}
\examples{
  data("nancycats", package = "adegenet")
  df <- adegenet::genind2df(nancycats)
  pop <- as.vector(df$pop)
  gen <- adegenet::df2genind(df[2:ncol(df)], pop = pop, ncode = 3)
  catpop <- adegenet::genind2genpop(gen)
  result <- calc_dplus(catpop)
  print(result)
}
