% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_KBA.R
\name{identify_KBA}
\alias{identify_KBA}
\title{Identify KBA(s) with distinct genetic diversity}
\usage{
identify_KBA(genpop, criterium)
}
\arguments{
\item{genpop}{A genpop object (from the adegenet package) containing an additional row "pop" containing the KBA in which the individual occurs.}

\item{criterium}{A string specifying the criterion to apply (A1: Threatened species (A1a: Critically endangered/ Endangered species, A1b: Vulnerable species, A1c: Critically endangered/ Endangered species due only to past/current decline \link{Red List A only, but not A3 only}, A1d: Vulnerable species due only to past/current decline \link{Red List A only, but not A3 only}), B1: Individual geographically restricted species)}
}
\value{
A vector containing the names of the identified KBA(s). The vector is empty if no KBA meets the criterium (character(0)).
}
\description{
This function identifies KBA with distinct genetic diversity for different criteria
}
\examples{
  data(nancycats, package = "adegenet")
  df <- adegenet::genind2df(nancycats)
  pop <- as.vector(df$pop)
  gen <- adegenet::df2genind(df[2:ncol(df)], pop = pop, ncode = 3)
  catpop <- adegenet::genind2genpop(gen)
  result <- identify_KBA(catpop, "A1a")
  print(result)

}
